+++ Netkan/Transformers/AvcTransformer.cs
@@ -19,6 +19,8 @@ namespace CKAN.NetKAN.Transformers
         private readonly IHttpService _http;
         private readonly IModuleService _moduleService;
 
+        public string Name { get { return "avc"; } }
+
         public AvcTransformer(IHttpService http, IModuleService moduleService)
         {
             _http = http;