+++ Netkan/Transformers/NetkanTransformer.cs
@@ -14,6 +14,8 @@ namespace CKAN.NetKAN.Transformers
     {
         private readonly List<ITransformer> _transformers;
 
+        public string Name { get { return "netkan"; } }
+
         public NetkanTransformer(
             IHttpService http,
             IFileService fileService,
@@ -22,7 +24,7 @@ namespace CKAN.NetKAN.Transformers
             bool prerelease
         )
         {
-            _transformers = new List<ITransformer>
+            _transformers = InjectVersionedOverrideTransformers(new List<ITransformer>
             {
                 new MetaNetkanTransformer(http),
                 new SpacedockTransformer(new SpacedockApi(http)),
@@ -34,13 +36,15 @@ namespace CKAN.NetKAN.Transformers
                 new VersionEditTransformer(),
                 new ForcedVTransformer(),
                 new EpochTransformer(),
-                new VersionedOverrideTransformer(),
+                // This is the "default" VersionedOverrideTransformer for compatability with overrides that don't
+                // specify a before or after property.
+                new VersionedOverrideTransformer(before: new string[] { null }, after: new string[] { null }),
                 new DownloadSizeTransformer(http, fileService),
                 new GeneratedByTransformer(),
                 new OptimusPrimeTransformer(),
                 new StripNetkanMetadataTransformer(),
                 new PropertySortTransformer()
-            };
+            });
         }
 
         public Metadata Transform(Metadata metadata)
@@ -50,6 +54,43 @@ namespace CKAN.NetKAN.Transformers
                     metadata,
                     (transformedMetadata, transformer) => transformer.Transform(transformedMetadata)
                 );
+        }
+
+        private static List<ITransformer> InjectVersionedOverrideTransformers(List<ITransformer> transformers)
+        {
+            var result = new List<ITransformer>();
+
+            for (var i = 0; i < transformers.Count; i++)
+            {
+                var before = new List<string>();
+                var after = new List<string>();
+
+                before.Add(transformers[i].Name);
+
+                if (i - 1 >= 0)
+                    after.Add(transformers[i - 1].Name);
+
+                result.Add(new VersionedOverrideTransformer(before, after));
+                result.Add(transformers[i]);
+            }
+
+            if (result.Any())
+            {
+                var firstVersionedOverride = result.First() as VersionedOverrideTransformer;
+
+                if (firstVersionedOverride != null)
+                {
+                    firstVersionedOverride.AddBefore("$all");
+                    firstVersionedOverride.AddAfter("$none");
+                }
+
+                result.Add(new VersionedOverrideTransformer(
+                    new[] { "$none" },
+                    new[] { result.Last().Name, "$all" }
+                ));
+            }
+
+            return result;
         }
     }
 }